requirements.txt
Flask==3.0.0
pandas==2.2.2
openpyxl==3.1.5
Pillow==10.4.0
